prodash not rendering

+++ glit-cli/src/global_option_handler.rs
@@ -5,10 +5,9 @@ pub struct GlobalOptionHandler();
 
 impl GlobalOptionHandler {
     pub fn config(matches: &ArgMatches) -> GlobalConfig {
-        // TODO: get real  number of thread from hardware
         let thread_num = matches
             .get_one::<usize>("thread")
-            .unwrap_or(&(8 as usize))
+            .unwrap_or(&(8_usize))
             .to_owned();
 
         let output = matches
